add ddd to api
