--- TelegramBotApi/Api/Controllers/Auth/Dto/Requests/LoginRequestDto.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Api.controllers.Dto.Requests;
-
-public class LoginRequestDto
-{
-    [Required] 
-    public string Login { get; set; }
-
-    [Required] 
-    public string Password { get; set; }
-}
