+++ TelegramBotApi/Api/DependencyRegistration/AddDomainDependencies.cs
@@ -1,18 +1,15 @@
 ﻿using Logic.Services;
 using Logic.Services.Interfaces;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace Logic;
+namespace Api.DependencyRegistration;
 
-public static class ForStartup
+public static class AddDomainServices
 {
-    public static IServiceCollection AddLogicServices(this IServiceCollection services)
+    public static void AddLogicServices(this IServiceCollection services)
     {
         services
             .AddTransient<IAuthService, AuthService>()
             .TryAddScoped<HttpClient>();
-
-        return services;
     }
 }
