+++ TelegramBotApi/Domain/Extensions/HttpRequestMessageExtensions.cs
