+++ TelegramBotApi/Domain/Services/AuthService.cs
