+++ TelegramBotApi/Domain/Services/Interfaces/IAuthService.cs
