+++ TelegramBotApi/Domain/Settings/AuthUrfuSettings.cs
