--- TelegramBotApi/Logic/Logic.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <Reference Include="Microsoft.AspNetCore">
-        <HintPath>C:\Users\misha\.dotnet\shared\Microsoft.AspNetCore.App\6.0.10\Microsoft.AspNetCore.dll</HintPath>
-      </Reference>
-      <Reference Include="Microsoft.Extensions.Configuration" />
-      <Reference Include="Microsoft.Extensions.Configuration.Abstractions">
-        <HintPath>C:\Users\misha\.dotnet\shared\Microsoft.AspNetCore.App\6.0.10\Microsoft.Extensions.Configuration.Abstractions.dll</HintPath>
-      </Reference>
-      <Reference Include="Microsoft.Extensions.DependencyInjection.Abstractions">
-        <HintPath>C:\Users\misha\.dotnet\shared\Microsoft.AspNetCore.App\6.0.10\Microsoft.Extensions.DependencyInjection.Abstractions.dll</HintPath>
-      </Reference>
-      <Reference Include="Microsoft.Extensions.Options">
-        <HintPath>C:\Users\misha\.dotnet\shared\Microsoft.AspNetCore.App\6.0.10\Microsoft.Extensions.Options.dll</HintPath>
-      </Reference>
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="6.0.12" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.10.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Dal\Dal.csproj" />
-    </ItemGroup>
-
-</Project>
