--- TelegramBotApi/TelegramBotApi.Tests/UnitTest1.cs
@@ -1,15 +1,0 @@
-namespace TelegramBotApi.Tests;
-
-public class Tests
-{
-    [SetUp]
-    public void Setup()
-    {
-    }
-
-    [Test]
-    public void Test1()
-    {
-        Assert.Pass();
-    }
-}
