+++ TelegramBotApi/TelegramBotApi.sln
@@ -2,17 +2,19 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{4F3A4F5C-A359-476E-8686-3C6C8CDD39C8}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Logic", "Logic", "{8BA09D31-FBDE-4DBC-AE5C-86FB191EC6D4}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{8BA09D31-FBDE-4DBC-AE5C-86FB191EC6D4}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Dal", "Dal", "{B16D2190-D01F-43D1-9EB2-1EFAC8C25A78}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dal", "Dal\Dal.csproj", "{403CE003-A8A6-4111-9BCC-2AA0421678B9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Logic", "Logic\Logic.csproj", "{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TelegramBotApi.Tests", "TelegramBotApi.Tests", "{2F8324A3-B5A1-476A-9F50-E1DBCEF7D0DC}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Application", "Application", "{0CD4815B-2EB2-4A62-ABDE-F2172A470714}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TelegramBotApi.Tests", "TelegramBotApi.Tests\TelegramBotApi.Tests.csproj", "{72D124B9-4E9D-4EC4-BACD-F1259F103103}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "Application\Application.csproj", "{8866A691-6F85-49A3-84AF-9238B7F68915}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Api", "Api", "{168665D8-77B4-43A9-A9D8-35589FC76D27}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,14 +34,15 @@ Global
 		{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565}.Release|Any CPU.Build.0 = Release|Any CPU
-		{72D124B9-4E9D-4EC4-BACD-F1259F103103}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72D124B9-4E9D-4EC4-BACD-F1259F103103}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72D124B9-4E9D-4EC4-BACD-F1259F103103}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72D124B9-4E9D-4EC4-BACD-F1259F103103}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8866A691-6F85-49A3-84AF-9238B7F68915}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8866A691-6F85-49A3-84AF-9238B7F68915}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8866A691-6F85-49A3-84AF-9238B7F68915}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8866A691-6F85-49A3-84AF-9238B7F68915}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{403CE003-A8A6-4111-9BCC-2AA0421678B9} = {B16D2190-D01F-43D1-9EB2-1EFAC8C25A78}
 		{2994DAE0-D1EF-450A-9D76-EC5D2B4A4565} = {8BA09D31-FBDE-4DBC-AE5C-86FB191EC6D4}
-		{72D124B9-4E9D-4EC4-BACD-F1259F103103} = {2F8324A3-B5A1-476A-9F50-E1DBCEF7D0DC}
+		{8866A691-6F85-49A3-84AF-9238B7F68915} = {0CD4815B-2EB2-4A62-ABDE-F2172A470714}
+		{4F3A4F5C-A359-476E-8686-3C6C8CDD39C8} = {168665D8-77B4-43A9-A9D8-35589FC76D27}
 	EndGlobalSection
 EndGlobal
